+++ app/[lang]/(main)/MainLayoutClient.tsx
@@ -7,9 +7,6 @@ import { Popup } from '@/types/interfaces';
 import Header from '@/components/layouts/Header';
 import Footer from '@/components/layouts/Footer';
 import { PicnicMenu } from '@/components/client/common/PicnicMenu';
-import PopupBanner from '@/components/client/vote/dialogs/PopupBanner';
-import { getLocalizedString } from '@/utils/api/strings';
-import { getCdnImageUrl } from '@/utils/api/image';
 
 const fetcher = (url: string) => fetch(url).then(res => res.json());
 
@@ -17,64 +14,21 @@ interface MainLayoutClientProps {
   children: React.ReactNode;
 }
 
-interface PopupSlide {
-  imageUrl: string;
-  title: string;
-  content: string;
-  popupKey: number;
-}
-
 const MainLayoutClient = ({ children }: MainLayoutClientProps) => {
   const pathname = usePathname();
   const { data: popups, error: popupsError } = useSWR<Popup[]>('/api/popups', fetcher);
-  const [popupSlides, setPopupSlides] = useState<PopupSlide[]>([]);
-  const [isPopupOpen, setIsPopupOpen] = useState(false);
-  const [currentSlide] = useState(0);
+  const [activePopup, setActivePopup] = useState<Popup | null>(null);
 
   useEffect(() => {
-    if (!popups || popups.length === 0) return;
-
-    const now = new Date();
-    const filtered = popups.filter((popup) => {
-      const hideUntil =
-        typeof window !== 'undefined'
-          ? localStorage.getItem(`hide_popup_${popup.id}`)
-          : null;
-      if (hideUntil && new Date(hideUntil) > now) return false;
-
-      const platform = popup.platform || 'all';
-      if (!(platform === 'all' || platform === 'web')) return false;
-
-      return true;
-    });
-
-    setPopupSlides(
-      filtered.map((popup) => ({
-        imageUrl: getCdnImageUrl(getLocalizedString(popup.image)),
-        title: getLocalizedString(popup.title),
-        content: getLocalizedString(popup.content),
-        popupKey: popup.id,
-      })),
-    );
+    if (popups && popups.length > 0) {
+      setActivePopup(popups[0]);
+    }
   }, [popups]);
 
-  useEffect(() => {
-    if (popupSlides.length > 0) setIsPopupOpen(true);
-  }, [popupSlides]);
-
-  const handleClosePopup = () => setIsPopupOpen(false);
-
-  const handleCloseFor7Days = () => {
-    const slide = popupSlides[currentSlide];
-    if (!slide) return;
-    const hideUntil = new Date();
-    hideUntil.setDate(hideUntil.getDate() + 7);
-    if (typeof window !== 'undefined') {
-      localStorage.setItem(`hide_popup_${slide.popupKey}`, hideUntil.toISOString());
-    }
-    setIsPopupOpen(false);
+  const handleClosePopup = () => {
+    setActivePopup(null);
   };
-
+  
   if (popupsError) console.error('Failed to load popups', popupsError);
 
   // Firebase Analytics: mypage 섹션 포함 전역 page_view 로깅
@@ -128,13 +82,6 @@ const MainLayoutClient = ({ children }: MainLayoutClientProps) => {
       <div className="container mx-auto px-4">
         <Footer />
       </div>
-
-      <PopupBanner
-        isOpen={isPopupOpen}
-        onClose={handleClosePopup}
-        onCloseFor7Days={handleCloseFor7Days}
-        slides={popupSlides}
-      />
     </div>
   );
 };
